requirements.txt
anthropic
langgraph
langchain-core
typer[all]
sqlalchemy
alembic
python-jobspy
playwright
google-search-results
python-docx
pypdf
apscheduler
google-auth
google-auth-oauthlib
google-api-python-client
requests
python-dotenv
rich
pydantic
pydantic-settings
pytest
pytest-asyncio
